Enables operating over empty files.
Both the `decrypt_file_and_save` and `write_file` functions don't work
with empty files because it is illegal to create an buffer of length
zero.  As a workaround, check if the file being edited does not exist,
and if so, assume that the plaintext is an empty string.  Similarly, if
the decrypted text is empty, then create an empty plaintext file.